LayoutTests/batterystatus/updates-expected.txt

Tests that updates to the battery event causes new events to fire.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS battery.charging is charging
PASS battery.chargingTime is chargingTime
PASS battery.dischargingTime is dischargingTime
PASS battery.level is level
PASS battery.charging is charging
PASS battery.chargingTime is chargingTime
PASS battery.dischargingTime is dischargingTime
PASS battery.level is level
PASS successfullyParsed is true

TEST COMPLETE
